Add support for multiple modules
Fix standard midi map: standard midi map is used by the GUI to communicate with the engine
and it is defined in utils. Engine can add a root module (the one that will play). When a root or
non root module is added a buf_in and buf_out can be specified.